Added main.cpp
Created in 2015.
Checks for all prime numbers between 3 and 100,000.
Uses the mod function 
i.e. 
if (i mod j ==0) {flag = 1} not prime
if (flag == 0) { cout << i is prime
Starts from 3 and skips multiples of 2
starts calculating the time taken to complete the process just before the loop and as the loop ends.